Updates drivers/tty/serial/imx.c to revision cc568849370bb131d896f4c5933cc72bf7ee603d
from git@github.com:torvalds/linux.git, which is somewhere in the 4.5 - 4.8 kernel series

This revision was selected based on the following criteria:
* recent enough that it has many fixes to DMA functionality
* is the revision *before* additional api dependencies are added (mctrl_gpio), which would make patching/integration more complex than it already is.

This revision does not build cleanly against linux-imx-3.10.17 due to api changes, but is applied independently to clarify which changes are from upstream and which changes are local to this project. edits to restore compatibility are introduced by a subsequent commit.